+++ src/App.tsx
@@ -5,7 +5,7 @@ import Layout from './pages/Layout';
 
 import Home from './pages/Home';
 import CampDetails from './pages/CampDetail';
-import CampApply from './pages/CampApply/CampApply';
+import CampApply from './pages/CampApply';
 import Community from './pages/Community';
 
 import NotFound from './pages/Error';
